[netsatck3] Support configuring transport protocols at compile time
This change introduces the `IpTransportContext` trait, which defines
the API that a transport-layer protocol implementation must provide
to the IP layer.

Based on that, we introduce the `Ipv4TransportLayerContext` and
`Ipv6TransportLayerContext` traits, which each provide one associated
type with an `IpTransportContext` bound for each IPv4 Protocol
number/IPv6 Next Hop number that a user might want to provide an
implementation for. Any type assigned `()` will result in the
associated protocol being unsupported (an ICMP "protocol unsupported"
error message will be generated for any incoming traffic on that
protocol number).

Note that not all protocol numbers may be configured. Protocols
that are used by the stack for internal operation are reserved.
Examples include ICMP, ICMPv6, IGMP, and a number of IPv6 extension
header values.

Test: No behavior change

Change-Id: Id0e892d30ca2e55ee14e00e0470280fbc7905b5d