[bt][async-helpers] Disallow multiple outstanding requests from a client
Multiple outstanding hanging get requests from a client over a single
channel should result in an error rather than all responses being
responded to. It is very likely that having multiple concurrent requests
is a bug and it can introduce ambiguity or different parts of a system
having incomplete pictures of the state of the world.

Change-Id: I792bb3649c7996c7421a02db48dd33a39ae241d7