[build][fuzzing] Allow fuzzer output_names, package_names.
This CL allows fuzzer authors to specify non-default names for fuzzers
and fuzzer packages. It disentangles GN names and labels, and provides a
cleaner "normalization" of fuzzers to achieve this.

It also simplifies the template by removing unused parameters (like
`sanitizers`) and merging overlapping fields (like `fuzzers` and
`fuzzer_profiles`).

Bug: 44955
Change-Id: I0cb5374ac52002720f800e88afba8673cc4ee26f