[component_manager] Update the Vmex service to work with the inbuilt root_resource
This change updates the fuchsia.process.Vmex service to use the
refactored root_resource_handle. It no longer needs to connect to
a service to get this handle which greatly reduces the complexity
of the builtin service. Since this will always be enabled now I removed
the builtin flag requirements as well.

Note: Moved this over to my open source account benwright@chromium.org.

Fixed: 47397
Change-Id: I3e1b50b1ec506fe94f2894fc74fb8a08654e2654