[runtests] Support for profile merging
Currently, we store every profile as a separate file, but that leads to
a dramatic output image size growth as these profiles tend to be fairly
large. However, many of these profiles come from the same shared object
(e.g. libc.so is the most common one) and the only difference between
different runs are the counters. We could dramatically decrease the size
by merging counters for profiles that originated from the same module,
thereby needing only a single profile per-module. This is already
supported by the profile runtime on other platforms, where the runtime
re-opens the existing profile if it already exists and updated the
counters instead of creating a new profile. This solution is not
applicable to Fuchsia since runtime on our platform doesn't have a
access to the filesystem, rather it publishes the profile as VMO through
the fuchsia.debugdata.DebugData service. This service is provided to
each test by runtests, which is also responsible for writing these files
to the output directory. We could therefore perform the profile merging
in runtests. This solution is implemented by this change.

Test: runtests-utils
Bug: 38110
Change-Id: I7339ec1baa854944beb42913e59f98b938de4d38