[graphics][compute][spinel] Return number of in-flight queue submissions
The `spn_vk_context_wait()` function now yields the number of in-flight
queue submissions.

This is critical for integrating with a fence-less swapchain.

If the in-flight queue submissions are 0 then we know that we can block
on the vkAcquireNextImage().

Otherwise, Spinel's dispatcher should be given a chance to drain.

Testability: There is no change in behavior -- a previously private
counter is being returned.

Change-Id: Ic66f6df6f8830ceaa6bc398cba7d7a0b38a1ca31