[SetUI][Core] Remove block_on from GenerateService.
GenerateService is used to inject services for test purposes. To allow
access to shared mappings of test services, we were blocking
whenever connecting to a service. This was also in alignment with the
existing FIDL connect API.

This change moves connect in ServiceContext to be an asynchronous
method. GenerateService has been updated to now return a
BoxFuture to enable asynchronous behavior in the closures.

Bug: 48120
Test: fx run-test setui_service_tests --  --test
Change-Id: I6153cae5a1adccdd75c12213e39a7da8aeddcc0c