[scripts] remove start-dhcp-server
This script sets configurations that are not safe for many common
environments and users have a tendency to run it blindly without
understanding of the consequences. Much of the unwanted behavior of
the script is unnecessary and should one day be replaced by a script
that is more fit for purpose and does not introduce these problems.

Change-Id: Iebb769581d8dfcae3bcc4bab1fda61bbfa548d6a