[components][cml] Disallow strong dependency cycles
Add checks to cmc and cm_fidl_validator to fail validation if a
dependency cycle is detected.

Two components have a dependency if one offers a capability to another,
assuming the dependency is not explicitly marked as weak.

Also update exitsing CMLs to mark weak dependencies where there are
currently cycles.

Change-Id: I3023bf22f683f79d901d7c3245284a5ef2541966