[sysmgr] Removes redundant configuration
The declarations of fuchsia.memory.Monitor and
fuchsia.memorypressure.Provider are already available
in services configuration for sysmgr.

This change removes the redundant declaration, allowing us to add
sticter checks that will fail sysmgr startup if further redundant
declarations exist.  This is intended to remove an ambiguity resulting
in duplicate "services" keys across many sysmgr configurations.

As the config fragment was accidentally included through a package
dependency, it seemed that the "startup_services" declaration was not
intended to be used in general, so is now omitted.

Bug: 39576
Change-Id: I25304677109e6ca4a807de557a4d83354ddeaaef