[FDF] Add logs to suspend path
Recently there have been cases on the field, where the filesystem
shutdown is done but the suspend taks fail to get created and there is
no sign of reboot afterwards. What is causing this is unclear.
This change does 2 things. 1.Add logs to the suspend path for better
debugging. 2. Reorders the checks for another suspend and moves the watchdog
creation before the suspend task creation immediately after filesystem shutdown.

Bug:47931, buganizer-150294009

Test: fx run-tests driver_manager_tests.

Change-Id: I7a2a686098baa574fa6904ce177f0c1c8f501da7