[sysmgr] Crash on inconsistent services config
Makes `sysmgr` exit with an error when parsing configuration files that
have duplicated service names as keys.  Since file processing order is
not defined, the results of parsing a file set containing conflicting
entries in the `services` map is unpredictable. And since service
configuration is defined on a product level, we should never have
conflicting service configurations.

This is an error (same `service_name`, different implementation across
files):

```
file1.json:
{
  "services": { "service_name": "service_impl" }
}
file2.json:
{
  "services": { "service_name": "service_impl_2" }
}
```

and this is an error too (same `service_name`, different implementation
in the same file):

```
file.json:
{
  "services": {
    "service_name": "service_impl",
    "service_name": "service_impl_2"
    }
}
```

Bug: 39576
Change-Id: I03b2d89fef69a4d1aa71fa38564586989296c9e7