[system-tests] Switch connection listeners to channels
Now that we are using contexts to implement timeouts, using sync.WaitGroup
caused us to spawn another goroutine in order to timeout waiting for the
wait group to finish. This switches sshclient to just use closing a channel
as that signal, since we can select on it being closed at the same time
as the context finishing.

Test: Manually confirmed OTA tests still work.
Change-Id: Ic893c52a155c805dd48c794215eb05b1a740b8f2