[paver] Avoid ever-increasing kernel priorities in Chrome OS.
Chrome OS uses a "kernel priority" flag in its partition table. The
bootloader by default will boot the highest priority kernel (with some
additional logic not relevant here).

When writing out Zircon partitions, we ensure that the ZIRCON-A
partition is set to "highest_priority + 1". Unfortunately, in this
calculation, we fail to take into account that our partition might
already be the highest priority, and will increase it any way. After 15
paving attempts, we run out of bits in the priority field, and paving
fails.

Update the logic to avoid taking into account our own partition's
priority when setting the new priority number.

Change-Id: I6fdadd80ddf5b4eba497a9bbbfc5ce75da6a3359