[fs][inspector] Remove heavy caching from minfs-inspector.
MinfsInspector would cache the entire inode table, bitmap
table, and journal in memory based on the superblock on load. This
forced applications to make sure the arguments they pass in are
within cache vmo bounds, making it impossible for users to ask
for arbitrary indices in the case of corruptions. Furthermore, it used
a lot of memory. Changed MinfsInspector api to only cache the superblock
and use a small single block vmo buffer.

Bug: 37907, 47554
Change-Id: Ifa5a3894715ff9706e760fe0872c6b161ee65863