[system-tests] Stop the server when shutting down
In order to shut down the server if the context timeout happens, we added
another goroutine that waited for the context to finish. However, if the
caller calls `server.Shutdown()`, the goroutine keeps running until the
context is canceled. To release these resources sooner, this patch adds
another `shuttingDown` channel that we'll close in `Shutdown()`. Then
this goroutine can also select on this channel to shut down early.

Test: Manually verified OTA tests still work
Change-Id: Ibf36647783bee67a2fa1900f62ba4865157c9aec