[SetUI] Remove block_on from Switchboard and Registry.
Switchboard and Registry currently utilize block_on during
creation in order to avoid asynchronous waits during bringup
in main. This leads to indefinite blocking if either decides to
do other asynchronous operations, such as waiting on a
receiver.

This changelist removes block_on in favor of running the
components' initialization on the executor.

Bug: 48120
Test: fx run-test setui_service_tests --  --test
Change-Id: Ia6a6de1da84091a9035c1d24b2612b952c901a5b