[lowpan] Initial version of LoWPAN Service and Dummy Driver
The LoWPAN service is a hub where LoWPAN devices can register
themselves. It enables applications to discover and use instances
of the LoWPAN Device protocol.

This commit also adds a new library for code that is shared across
LoWPAN interface drivers written in Rust. It also includes an
example "dummy" driver which is used for the included integration
test.

Test: fx run-test lowpan-tests
Change-Id: I1bd2e5448df75b96ab337e8f569eb85a3972dc28