[wlanif] Release mutex before calling driver routine.
Fixed: 42095

The assoc response might be sent back in the assoc request call
stack itself (in case of a failure in setting it up). Release the
mutx before calling the driver's assoc req routine.

Test: Forced a failure in the driver assoc request sequence and
wlanif does not hang anymore.

Added a unit test to verify association request is handled
correctly

Change-Id: I8b53f4190017d919e994f932880e9c4f43b8e737