[debugger] Remove assertion for null CFA.
There was an over-eager assertion that expected a non-null frame
fingerprint when stepping. This implies a non-null CFA (Canonical Frame
Address -- computed by the unwinder). Go code seems to cause the
unwinder some difficulty (a separate problem) and in Go the CFA can be
0.

This can also theoretically happen in unsymbolized code so is a case we
need to handle even if the unwinder worked for Go in all cases.

This patch just removes the assertion. The code does the best it can
already in this situation, and it seems to step as expected for Go.

Bug: 48100
Change-Id: I03385756190ac850e2fe081f8ac340a101bcb863