[debugger] Add special identifiers
Implements the first part of unified handling of special identifiers
from the design doc in the linked bug. This attempts to be a unified way
to name special things like the program entrypoint, to annotate
identifiers like PLT functions and register names, and to escape
otherwise-unparseable identifier names in the DWARF symbols.

This current patch supports "$main" for the main entrypoint, "$anon" for
anonymous namespaces, "$plt(...)" for PLT symbols, and "$reg(...)" for
register names (unannotated register names are still supported if there
are no collisions with variables).

The old PLT symbol syntax is still supported so we can transition users
in a separate phase.

Bug: 5722

Change-Id: Iea286a49b5233613d897700eec9741ad2f8e9f39