[kernel] Panic if OOM reboot takes longer than 10 seconds
Add a panic deadline to platform_graceful_halt_helper and pass it to
the underlying calls that might block for an extended period.  If any
of the calls fail for any reason (including timeout),
platform_graceful_halt_helper will panic the system.

platform_graceful_halt_helper is "upstream" of the crash log so if we
call platform_graceful_halt_helper with reason OOM, but end up
panicking, the reported reason will show panic.

The purpose of this change is to 1) ensure that when a system OOMs, it
will eventually reboot and 2) produce evidence that helps us track
down why the graceful reboot did not complete.

Calls to platform_graceful_halt_helper outside of OOM are not
affected.

Tested by inserting an infinite loop for a single secondary CPU just
before its call to arch_flush_state_and_halt.  Triggered an OOM via "k
pmm oom".  Observed that the deadline was exceeded, the system
panicked, and the crash reason was correctly reported as panic instead
of OOM.

Bug: 48076
Change-Id: Ib49f11b99369b0b983003521eba85be9f9a23c2f