[kernel][vm][arm64] Protect non-ram regions of physmap
Protect the memory regions of the physmap that are not backed by RAM
to prevent hardware based prefetching from inadvertantly reading
device memory.

Note, we are not doing anything to the non-ram regions on x86 because
the kernel has dependencies on some non-ram regions of the
physmap (e.g. smbios).

Bug: 46799
Test: booted qemu/cavium, astro, and sherlock; ran through GI
Change-Id: Iea326b8e1e4752d7d6728c55a218cd3297e1c5ca