[pkgfs] Log on loading non-allowlisted dynamic packages, add an optional enforcement mode
Add a pkgfs arg to not return these packages ("enforcement mode"),
used for testing and to be turned on later.

With this change, pkgfs will only complain when dynamic packages are
accessed through /pkgfs/packages. Access through /pkgfs/versions is
unchanged.

This works by parsing an allowlist from the system image, loading it
at pkgfs startup, and logging if any non-static packages are opened or
read which aren't on that allowlist. Enforcement of these rules is
toggled by a pkgfs command line arg.

Bug: 44522
Change-Id: I410ae421966d5d90df9bfde1387e9c22379e0448