[recovery] add an FDR-only variant
Includes some fixes for assemble_system and dependency templates to
ensure they produce unique labels and outputs.

The current recovery ZBI is renamed to recovery-eng, and a new
recovery-fdr is introduced.

Both images are now built into target_out_dir instead of
root_build_dir, and the test is updated with the new file location.

Bug: 48237
Change-Id: Ib220ea8febf87473dbf0e0468c93c5138104e027