[SetUI][core] Move messenger creation to separate task.
Currently, messengers are created from a shared message hub
instance. This pattern requires acquiring a lock on the message
hub before accessing the hub, complicating task ordering. For
example, a deletion op of a messenger would require acquiring a
lock, which is not possible without spawning a new thread in a
synchronous method like drop from the Drop trait.

This changelist addresses this issue by pushing messenger
creation in the MessageHub to its own task. All requests are
queued into a channel received by this task to provide
ordering. Also, the MessageHub is no longer provided to
components. Instead a MessengerFactory is handed back.
Since the MessengerFactory communicates with the
MessageHub via the previously mentioned channel, the
factory can be cloned and interacted without acquiring
a lock.

This changelist also enforces unique addresses for
messengers and will return an error on messenger
creation if this criteria is not met.

Fixed: 48120
Test: fx run-test setui_service_tests --  --test
Change-Id: I89e3e0843e6eafe0f066e5490d8efe80cc3003f8