[kernel][object] Assert KernelHandle holds only ref to Dispatcher
KernelHandle is used to ensure a Dispatcher's on_zero_handles method
is called in the event that the system could not create a Handle to
hold it (e.g. handle table is full).

It is an error to construct a KernelHandle around a Dispatcher with
outstanding Handles.  Prior to this CL, doing so could result in a
double call to on_zero_handles.

This CL changes KernelHandle's value constructor to assert that the
Dispatcher's handle count is 0.

Change-Id: I8e0e1597b5d2b5b9d70c0dade7fd600e265c74f6