[board][x86][cros] Trivial: Rename "IssueCommand" parameter names.
Update the "EmbeddedController::IssueCommand" parameter names to avoid
using "output" to mean "what is output to the EC", and "input" to mean
"what is read from the EC".

Instead use the more common usage of "input" to mean "input to the
function" and "result" to mean "result of the function".

Change-Id: I1f7e96785f30062f3f6a50a07450b226c60e398f