[reland][bundles] Build and test Topaz for fuchsia.git commits
This CL causes changes to fuchsia.git to also build and test the code in
topaz, which means build and test failures will be caught earlier (e.g.,
before a CL lands rather than when trying to roll into Global
Integration).

This is effectively a reland of fxr/308987 which was reverted because it
caused fuchsia.git builders to take 10-15 minutes longer.  As most of
Topaz has been migrated, and Infra has become faster in the last 6
months, this change now causes a 2-3 minute timing regression in Infra.

Change-Id: I6c994ac867badaae891bbe1d6f7ca40bca6a76e1