[SetUI][setup] Add Power to all setting types
Power was left out of the list of all setting types, which is the
default set used to start up the environment. As a result,
changing the setup network interfaces could not reboot
and therefore failed. This changelist adds the setting type
back to the set.

Test: Existing tests verified behavior with the setting type
present and therefore passed. The client behaved correctly
in throwing an error with this setting missing. Work to allow
the set to be configurable by product is underway.
Fixed: 48211

Change-Id: Idf840f182f183cf61b49411edfcd23fad793c6de