[sysmgr] Validate sysmgr's configuration at build time
This adds a validation step to the config_package template to check that
sysmgr's configuration does not have conflicts which will result in
(currently) undefined behavior or (soon) crashes at runtime. A new
Python tool reads in the generated manifest of files that will form the
'config-data' package, parses each of the files that will contribute to
sysmgr's config-data, and checks for conflicts in the 'services' key.

Note to developers:

This commit will be validated through Global Integration before landing
it to ensure that there are no existing problems in supported build
configurations.

If this change appears to cause a build failure for you locally that
says "Error: conflicts detected in sysmgr configuration", you likely
have a problem in your 'fx set' or 'args.gn' that was causing subtle
failures before and need to adjust how you build. Please reach out
before reverting.

Bug: 39576
Fixed: 48223
Change-Id: Ib66d38760ed886d49f9111db8c8182a3213dc636