[memory_metrics] Log the Digests each time we send Metrics.
It would be useful to be able to see memory usage in the log
periodically when tracking down problems. We already collect Digests
every 10 minutes so the logging shouldn't add much cpu usage and by
logging all buckets on one line at 10 minute intervals it shouldn't spam
the logs excessively.

Test: None <only change is logging>
Change-Id: I3090374cfaf47e2ba6bf07d1ccd47fed99f837c9