[msd-arm-mali] Don't timeout if device request semaphore is signaled
We may be seeing cases where the device thread has work queued up, but
it is taking a long time to be scheduled and when it is scheduled is
marked as timed out. In that case check if the semaphore is signaled and
avoid actually marking it as timed out if it has more work it can do.

Change-Id: I3119a2245874f45d942bd120f09619635f33e83b