[SetUI][core] Handle messenger deletion.
MessageHub currently doesn't do anything with Messengers
that go out of scope. As a result, address space is not
reusable. This changelist addresses that by giving the client
components a shared ActionFuse that will signal the
Messenger's deletion when all holders go out of scope.

To accomplish this, a MessengerClient is now handed to
the client instead of the Messenger (similar to a
MessageClient). This change was necessary since
Messenger is held by internal components, such as the
associated Beacon. An additional MessengerAction for
delete has been added in order to synchronous
deletion with creation.

Fixed: 47899
Test: A new test has been added to ensure cleanup
when components to out of scope.
fx run-test setui_service_tests -- --test

Change-Id: Ib24ed24de5bf33443b5deda85f4198bacedb235f