[system-tests] Allow tests to OTA into zedboot
This adds a "--ota-to-recovery" flag to the upgrade test that OTAs to
the fuchsia-pkg://fuchsia.com/update-to-recovery update package, instead
of using `dm reboot-recovery`. This allows the OTA test to be used on
devices that don't support `dm reboot-recovery`, or have a recovery
partition.

Test: Manually confirmed OTA tests OTA into recovery, and the OTA test
successfully passes when this flag is set.

Fixed: 48139

Change-Id: I82f1b25697052daffed0c219f5ca3d04d1564a0d