[media][tests] clean up test
This test-only CL cleans up the logging in the recently-added clock
library unittests. It also adjusts the name of the resultant binary
and the source (and resultant) .cmx file, so that both of these use
underscores (to fly in formation with the rest of media/audio).
Note that the component name continues to use hyphens, as required.

Test: no production change, test only

Change-Id: I25e99c63674fa7f618e63ac05f1c282676a92538