[SetUI][core] Refactor handler context.
Currently contexts are shared between all handlers
created by the Registry. Upcoming changes require
per handler values to also be passed at creation time.
As a result, it will be necessary for Contexts to be
unique per client.

This changelist nests shared values previously in
Context in an Environment struct, a child member
of Context.

Bug: 48343
Test: fx run-test setui_service_tests -- --test
Change-Id: I289a2df00b04094c02e98f026f1ef1649c90cb70