[diagnostics][plat] Diagnostics trie iteration should iterate over all nodes, even if empty.
For node hierarchy, these empty nodes should still show up, and use a None value for
properties to encode that there were no entries.

Additionally, when filtering node hierarchies, as long as there are ANY relevant selectors
for a given node, the node is included in the filter, even if no properties match.

Nodes are still dropped if there are no valid selectors matching that node.

BUG=48098

Change-Id: Ic0fe8b1de2419b3f8e5f2181b03eb843aa499e17