[kernel][arm64][zbi] add support for multiple RAM sections from bootloader
Previously the logic would only accept the first ZBI RAM entry and then
subsequent passed in entries would modify the first one. Change the
logic to handle up to 16 ram sections as separate PMM arenas.

This also changes the logic from the qemu bootshim slightly to instead
of appending a new ZBI ram section that modifies the first one simply
only accept the ram entry from the FDT. Since we always parse the FDT in
the qemu boot shim we should be okay.

Tested: booted on qemu and CQ, validated that memory limit command line
seems to properly trim entries. Artificially passed in a multi-arena ZBI
and validate it finds it properly.

Change-Id: I3347c8aba9e8ef1b6cf8c0a4b97580acc6b75685