[fdio] Implement eventfd
This CL implements eventfds on Fuchsia. The current implementation is
in-process, which means we cannot transfer eventfds to another process.
If we need to support transfering to another process, we will need a
kernel primitive that can store the uint64_t value of the eventfd.

This CL also introduces fdio_hosttests, which cross-tests FDIO against
the host system.

Bug: 47651
Change-Id: Iac90131aff26af3858f9d308c10ad4be9de1ab42