[visalia][audio] Fix fifo_depth
We set the input driver fifo_depth to the size of 3 channels of PCM
input data DMA chunk (21.3msecs) times 2. The driver won't see any data
from the DMA before 21.3msecs, this is the first delay inherent to
the DMA configuration used. Once the DMA is completed, with fifo_depth
corresponding to 42.6msecs, the driver has up to 21.3msecs to decode
the PDM data and place it in the ring buffer.

Test: runtests -t syn-audio-in-test
Test: audio -c 3 record /tmp/t.wav 10, on Visalia.
Change-Id: Id82ae666bc8cc127579d27ac5996d8b694aa13bf