[audio_core] Reference clock APIs
This CL adds GetReferenceClock and SetReferenceClock methods
to the AudioRenderer and AudioCapturer protocols, without
full implementation.

For now, any custom clock provided by a client is only used
in AudioRenderer for 'ref_time' values provided in the Play
and Pause callbacks. Neither AudioRenderer nor AudioCapturer
use the custom clock for timing and rate-adjustment, at this
time.

Test: 12 new test cases in audio_fidl_tests
Bug: 13613, 46620, 46621, 46622, 46623, 46628

Change-Id: Ideff7896a11f908a7d05750c00274fb9936715cc