[bindc] Add depfile handling to the compiler
Depfiles are used by GN and other build systems to track the
dependencies of a target. We can't handle the dependencies in GN itself
because the bind build target uses metadata to collect the libraries
used by a bind rule target, and metadata can't feed into the
target's dependencies.

Note that we could remove libraries that aren't actually used during
compilation from the depfile, but I don't think it's worth the extra
complexity. Each bind library target contains a single source file so
the only way this optimisation would help is if the dependencies for a
target were overspecified.

Change-Id: I7ff95a846a3cd8ba3d4b2f7d3f649ce08afce876