[paver][vmo] Commit payload VMO before writing partition
On the OTA path, the payload VMO is a blob, which can be pager-backed
(once demand paging is enabled). Make sure that the payload VMO has its
pages committed in memory before attempting to write it to a partition.
Not doing so can cause deadlocks in the block server due to read faults.

Without demand paging (current behavior) this change commits the payload
VMO's pages a little earlier, instead of doing it via pinning from a
block server write. A blob's pages are populated on initialization; the
payload VMO is a clone of the blob VMO. The block server pins the
payload VMO for the write, which is equivalent to a COMMIT operation.

Test: Added data-sink-test.cc which tests DataSinkImpl::WriteAsset()
with a pager-backed VMO. No observed change in paver behavior without
demand paging (default today). Made sure we are using an equivalent
amount of memory with `k zx vmos`. Tested with blobfs user pager as
well to make sure the OTA path succeeds.

Bug: 48147
Change-Id: I26b46845dd556a89e5a537fec90289607cc1b483