[tools][zbi] Enable "long-distance matching" mode in zstd
This is what the `--long` switch to the `zstd` command-line tool
does.  It's now turned on by default in the `zbi` tool when the
compression level is the default or higher.  While this may increase
compression time and memory use to some degree, it reduces ZBI sizes
by 8%-13% (depending on the image composition and the compression
level set in the `zbi_compression` GN build argument) and may even
speed up decompression beyond the benefit of the size reduction.

Change-Id: I7c97ab3842ca5cc0cbd671128e90220bb77de792