[kernel][vm] Handle an edge case return code.
Handle an edge case return code which can come back from
copy_array_(to|from)_user_capture_faults.  This function will attempt
to move data to or from user memory, but if a fault occurs, it will
attempt to capture the nature of the fault (the fault VA and the fault
flags) so that calling code can manually deal with the fault if
needed.

There is an edge case, however.  If the user-mode address range exists
partially or totally outside of the possible user space address range,
then the copy routine will make no attempt to actually perform the
copy, and return ZX_ERR_INVALID_ARGS instead.  This is a security
measure meant to outright reject _any_ attempt to go outside of the
user space address range in an attempt access memory.

In this case, however, the returned fault virtual address and flags
are not filled out.  In the call sites addressed by this CL, the
storage for these variable are sitting on the stack as uninitialized
variables.  Since the copy routine returned an error, the code was
then passing these garbage variables to the VmAspace::SoftFault
handler in an attempt to actually handle the fault.  If the SoftFault
handler thought that it actually successfully handled the fault (which
it could have, esp if the address in the fault VA was a legal user
space address), then SHOULD_WAIT will be returned to
ReadWriteInternalLocked, which will take this as a sign that it should
try again since there was a fault, but the fault has now been handled.

From here on out, we are pretty stuck.  We keep hitting the same stack
location for the fault address, and keep getting the same garbage
fault VA.  SoftFault/PageFault thinks it has taken care of the
situation for us, so we try again, and again, and again forever.

So, instead of allowing this, handle the special case of INVALID_ARGS.
Translate the result to NOT_FOUND instead as this is what PageFault
would have told us if we tried to access a page for which we had no
mapping in our address space.

Fixed: 46073
Change-Id: I521d2bcec74756bc9b7248b4d06dbf221174cf3e