[artifactory] Stop compressing blobs
This is blocking the ephemerality pipeline (as compressed blobs are not served from GCS with Content-Length set, a requirement of the receiving pkg-resolver).
Moreover, there is no longer a general wish have such objects compressed (e.g., on the part of nsylvain@).
I have left compression on for binaries and images alone.

Bug: 10427
Change-Id: I7777232ff2dea632085abcada8892a40ec9aa859