[audio_core] Add virtual audio input to admin tests
The audio_admin_test suite was using Loopback capturers to test admin
policy by setting the usage on the loopback capturer. We will shortly be
removing the ability to set a usage on a loopback capturer, so we need
to update these tests.

We do this by adding a new virtual input device that our non-loopback
capturers can pull data from, and we convert any capturer that needs to
have audio admin policy rules from a Loopback capturer to an Input
capturer.

Test: fx run-test audio_admin_tests
Change-Id: I1fec9730c996c7965b04daa9a3197bba50e1102d