[wlan] Update inspect log for privacy review
Main changes are:
- Remove BSS description when join scan fails due to incompatible BSS.
- In its place, add a "candidate_bss" node with BSSID and its hash. Only
  the hash will be included in the allow list.
- Change "connecting_to" from a property that includes SSID to a node
  that includes SSID and its hash. Only the latter will be in the allow
  list.
- Add hashes for SSID and BSSID under "connected_to" node.
- Scan failure in join scan events logged under "result" property for
  more uniformity.

Orthogonally, do a lossy UTF8 conversion for SSID when logging them in
Inspect instead of logging the error when it fails. This is because the
error will probably not be useful, and our hash should help us
differentiate two SSIDs even if they are not UTF8 valid.

Bug: 47502
Test:
1. fx run-test wlan-sme-tests
2. Manually verified Inspect log change

Change-Id: Iee30b4ed989b2bac629b50ac99ed8334f2baf782