[covargs] Don't close repo files until after they're used
In case of cloud repo files, these are downloaded into the cache and
hardlinked into a temporary location which is returned as the file path.
Upon closing the file, the hard link is deleted. Since llvm-cov needs
to use the file, we have to change the structure of the code to avoid
closing repo files until after they're used.

Change-Id: I8992216e7fb0b565d2d2d7a0123d0a6ba3f0ccdb