[trace] add TRACE_TRIGGER to provider library
This CL adds a TRACE_TRIGGER macro to the trace library.
TRACE_TRIGGER provides an out-of-band (with respect to trace events)
signal that can be mapped to actions such as stopping a
trace session. This CL addresses only the provider-side changes,
delivering triggers to the trace manager over the fifo used for
out-of-band communication.

For now, triggers are only sent after the session is started. This
means triggers cannot yet be used to start a session. Fixing this
involves some synchronization intracacies, and is left for a later
CL.

TEST: added a unit test

The unit test controls the internal 'session' object, mimicking what
happens when the trace manager initializes and starts a provider.
The provider protocol implementation has not been changed, so this
test provides good coverage for these changes.

Change-Id: I4473c0c633795c367f4611593aa7c1ae19613aaa