[libc][sanitizers] Fix false LSan report for __cxa_thread_atexit_impl
When a leak check is done (i.e. the process exits) while a thread
that previously used __cxa_thread_atexit_impl is alive, LSan
falsely reports the malloc allocations it did as leaks.  This is
because __sanitizer_memory_snapshot failed to include the pointer
that holds those allocations in its root set.

Bug: 45047
Change-Id: Ie5155e9e228d5b7b4e3464ca46348d39b05e2666