[cmdline] Add additional "const" to argv parameters.
Currently, calls to Parse() with a const array to const data such as the
following will fail:

  const std::vector<const char*> args = {"--help"};
  Parse(args.data());

This is because the declaration of "Parse" only declares the inner
string contents being const, and not the array itself.

Add an additional "const" to tell callers that the array itself also
will not be modified by the parser.

Change-Id: Ic9af349f7453b53440d065e777781587755cefb1