[audio][mixer] Disable position notification.
Disable position notification for now.  Currently, they are only used
for logging, but according to analysis done by Andres (see bug
151363928), something seems to be going wrong on Astro after enough
time, and the mixer stops reading all of the notifications.
Eventually, they build up and start to cause problems.

For now, turn them off as they are not (currently) needed for proper
operation.  We can turn them back on when we get to the bottom of bug
151363928.

Bug: 151363928
Change-Id: If90c916e1dbe720c1139041bbe0c0439c4ea93d9