[gpt] Fix memory leak in test
gpt_device_test provides a device list to be filled out by the
PartitionTable instance. The contents of the list were not deleted when
the PartitionTable is destroyed. Changed to vector of unique_ptr to
delete when test exits.

Reenabled LSan for GPT gpt_device_test.

Bugs: 47440
Tests: Verified memory leak is gone with LSan.
Change-Id: I0f30dd6158f6c2af97d011aeef352520db415c0a