[bt][l2cap-bredr] Fix use-after-free in OnRxConnRsp
In BrEdrDynamicChannel, use a WeakPtr to guard the OnRxConnRsp handler
for Connection Responses that may come after the DynamicChannel is
deleted.

The UAF may not actually be possible in the current system because
SignalingChannel and DynamicChannelRegistry are deleted in close
succession without handling commands, but it doesn't hurt to protect
against regression.

Bug: 924
Test: In bt-host-unittests,
L2CAP_BrEdrDynamicChannelTest.ChannelDeletedBeforeConnectionResponse

Change-Id: I95b5e55e5be796a15b7c8c292bf0c2240e16e2e8