[build][fuzzing] Remove rust-asan-fuzzer variant.
This CL merges the `rust-asan-fuzzer` back into `asan-fuzzer`. It was
initially split out because `rust-asan` could not be added to the `asan`
variant as binaries linked with rustc would fail. `asan-fuzzer` however,
builds no such binaries, and simplifies building.

While testing, a bug in rustc was observed for host fuzzers. Go and Rust
host fuzzers have been disabled by a common mechanism in this CL.

Bugs: 44551, 48624
Change-Id: I23dfcb4a56b26c578373cc7d2e08fb07c1d44c7a