[amlogic_decoder] Import H264 lib from chromium
This library will be used to manage the h.264 DPB. Only the essential
code is imported - dependencies on base/ and gfx/ will be replaced with FXL,
C++17 std, or small implementations with no dependencies. This is
because otherwise the nested web of dependencies would eventually force
us to pull in much more code than we really need.

Change-Id: I85bceacf976887b5549098d1723cb07d0fe55fea