[sl4f][benchmarks] Be stricter when processing infra environment vars
There was recently a regression where the infra recipes weren't
passing the env vars INPUT_COMMIT_HOST, INPUT_COMMIT_PROJECT and
INPUT_COMMIT_REF on the try builders.  That caused this code to stop
running catapult_converter on try builders, which meant there was some
amount of testing that was failing to happen pre-commit.

That regression was fixed here:
https://fuchsia-review.googlesource.com/c/infra/recipes/+/372873

Make convertResults() stricter so that it raises an error if some but
not all of the env vars are present.  That should prevent this testing
gap from occurring again.

Bug: 23091
Test: new check in unit test in performance_test.dart
 * run unit test locally with:
     fx set terminal.x64 --with sdk/testing/sl4f/client:tests
     fx run-host-tests sl4f_client_tests
Change-Id: Ifd5c981b244292f746ab243924a2799e48d630e5