[gn-sdk] Add GN rules to build a test workspace
This CL adds build rules to build the GN SDK. After this
change you will be able to build and test the GN SDK with by
running:

  1. `fx set core.x64 --with //scripts/sdk/gn:gn_sdk_test_workspace  --args="build_sdk_archives=true`
  2. `fx build`

The build rules pass through commands to run the
//scripts/sdk/gn/generate.py script to create the GN SDK as well as a
test workspace. After the build completes a GN SDK will be in
$OUT_DIR/sdk/gn/fuchsia-sdk directory and the test workspace will be
at $OUT_DIR/test_sdk/gn/test_workspace. Tests can then be run locally
by running the `run.py` script in the test workspace.

This CL was written to make it easier for contributors to build and (in
the future) test their changes to the GN SDK locally before submitting
their CL to CQ and/or for review.

Change-Id: Ib5fed16445b830cb088ee8a1186790a039e7f55f