[sl4f] Fix ProfileServerFacade service options
Due to the implementation of additional protocol
descriptors in SDP, this CL looks to fix the SL4F
api to match the inputs implementation.

This broke SdpSetupTest and BluetoothCmdLineTool
testing as there was a loose interpretation of
how the input was processed. BluetoothCmdLineTesting
for PTS was used to verify this CL works.

Bug: None
Test: BluetoothCmdLineTest
Change-Id: If30f778551d3e8f0987f294cb9d295d1747b47b2