[netstack3] Introduce DualStateContext, RngStateContext
This commit introduces a generalization of `StateContext` which
allows accessing two different states at the same time. This, in
turn, allows us to define `RngStateContext`, which is a
generalization of `RngContext + StateContext` which allows
accessing the RNG and the state at the same time. This allows us
to avoid creating temporary, mock RNGs when state access is
required at the same time.

Test: Existing unit tests exercise the RNGs; run with:
      fx run-test netstack3_tests -t netstack3_core_lib_test

Change-Id: I39b92426c055305bfa94e0b5a96d2ecedab30a94