[fidl] Convert C binding tests to packages
This involves making the following mechanical changes:
- move each zbi test out of zircon and into its own dir
  in src/lib/fidl/c
- each of these new dirs exposes a test executable target
- src/lib/fidl/c's BUILD file defines a test package that
  runs of the test executables, and defines cmx file for
  each

The new test package is called `fidl-c-test`. Dashes are
used instead of underscores as the use of underscores in
package names is being deprecated; eventually other FIDL
tests should be converted to use dashes.

One remaining test zbi test (`fidl-test`) is left to be
converted in a separate change, as it has a more
complicated GN rule

Test: fx test fidl-c-test
Change-Id: I575c4e02efabab1c296df94bbf1953e43930cd89